Extract useUsageBatching from App.tsx
Move the per-button click buffer, the 10s flush interval, the
beforeunload listener, and the unmount flush into
useUsageBatching(barId). Drop the now-unused increment import.

53/53 tests pass — including the existing 'batches button usage
updates' benchmark, which validates that 5 rapid clicks turn into
a single Firestore updateDoc with increment(5).
