+++ src/App.tsx
@@ -30,7 +30,6 @@ import {
   deleteDoc,  // Delete a document.
   arrayUnion, // Add elements to an array field.
   arrayRemove, // Remove elements from an array field.
-  increment,  // Atomic increment of a numeric field.
 } from 'firebase/firestore';
 // Import initialized Firebase instances and helper functions.
 import {
@@ -48,6 +47,7 @@ import { useLatestRelease } from './hooks/useLatestRelease';
 import { usePwaInstallPrompt } from './hooks/usePwaInstallPrompt';
 import { useGodMode } from './hooks/useGodMode';
 import { useMyBars } from './hooks/useMyBars';
+import { useUsageBatching } from './hooks/useUsageBatching';
 import { pMap } from './utils/async';
 
 
@@ -186,8 +186,6 @@ function App() {
   // Ref to track dragging state to prevent accidental clicks.
   const isDraggingRef = useRef(false);
 
-  // Ref to hold usage buffer for batching writes.
-  const usageBufferRef = useRef<Record<string, number>>({});
 
   // Dialog visibility states.
   const [showOffClockDialog, setShowOffClockDialog] = useState(false);
@@ -773,43 +771,8 @@ function App() {
     return btn.children || [];
   };
 
-  // Flush usage buffer to Firestore.
-  const flushUsage = useCallback(() => {
-    if (!barId) return;
-    const buffer = usageBufferRef.current;
-    if (Object.keys(buffer).length === 0) return;
-
-    // Reset buffer immediately.
-    usageBufferRef.current = {};
-
-    const updates: Record<string, any> = {};
-    for (const [btnId, count] of Object.entries(buffer)) {
-        updates[`buttonUsage.${btnId}`] = increment(count);
-    }
-
-    updateDoc(doc(db, 'bars', barId), updates)
-      .catch(e => console.error("Failed to flush usage", e));
-  }, [barId]);
-
-  // Periodic flush of usage stats.
-  useEffect(() => {
-    const interval = window.setInterval(flushUsage, 10000); // 10 seconds
-    // Also flush on page unload (refresh/close).
-    const onUnload = () => flushUsage();
-    window.addEventListener('beforeunload', onUnload);
-
-    return () => {
-        window.clearInterval(interval);
-        window.removeEventListener('beforeunload', onUnload);
-        flushUsage(); // Flush on unmount
-    };
-  }, [flushUsage]);
-
-  // Track usage for sorting.
-  const trackButtonUsage = (btnId: string) => {
-    if (!barId) return;
-    usageBufferRef.current[btnId] = (usageBufferRef.current[btnId] || 0) + 1;
-  };
+  // Buffered button-usage writes (10s flush interval).
+  const { trackButtonUsage } = useUsageBatching(barId);
 
   // Sort buttons based on custom order or usage.
   const sortButtons = (btns: ButtonConfig[], contextId: string) => {
